确认订单

+++ app/src/main/java/com/vip/uyux/model/OrderConfirmbefore.java
@@ -24,7 +24,7 @@ public class OrderConfirmbefore {
      */
 
     private int is_address;
-    private int sum;
+    private double sum;
     private AdBean ad;
     private String vipKey;
     private String vipDes;
@@ -43,11 +43,11 @@ public class OrderConfirmbefore {
         this.is_address = is_address;
     }
 
-    public int getSum() {
+    public double getSum() {
         return sum;
     }
 
-    public void setSum(int sum) {
+    public void setSum(double sum) {
         this.sum = sum;
     }
